enable always block option for external content,
close #3176
